fixed history for JS builds
