+++ pry/io_js.go
@@ -35,21 +35,21 @@ func readFile(path string) ([]byte, error) {
 	return ioutil.ReadAll(r)
 }
 
-type BrowserHistory struct {
+type browserHistory struct {
 	Records []string
 }
 
-// NewHistory constructs BrowserHistory instance
-func NewHistory() (*BrowserHistory, error) {
+// NewHistory constructs browserHistory instance
+func NewHistory() (*browserHistory, error) {
 
 	// FIXME:
 	// when localStorage is full, can be return an error
 
-	return &BrowserHistory{}, nil
+	return &browserHistory{}, nil
 }
 
 // Load unmarshal localStorage data into history's records
-func (bh *BrowserHistory) Load() error {
+func (bh *browserHistory) Load() error {
 	hist := js.Global().Get("localStorage").Get("history")
 	if hist.Type() == js.TypeUndefined {
 		return nil // nothing to unmarashal
@@ -64,7 +64,7 @@ func (bh *BrowserHistory) Load() error {
 }
 
 // Save saves marshaled history's records into localStorage
-func (bh BrowserHistory) Save() error {
+func (bh browserHistory) Save() error {
 	bytes, err := json.Marshal(bh.Records)
 	if err != nil {
 		return err
@@ -75,9 +75,9 @@ func (bh BrowserHistory) Save() error {
 }
 
 // Len returns amount of records in history
-func (bh BrowserHistory) Len() int { return len(bh.Records) }
+func (bh browserHistory) Len() int { return len(bh.Records) }
 
 // Add appends record into history's records
-func (bh *BrowserHistory) Add(record string) {
+func (bh *browserHistory) Add(record string) {
 	bh.Records = append(bh.Records, record)
 }
